Adding option to not propagate height/width to ReflexElement child
I was having an issue where I wanted the width of my components to be automatically updated, but not the height so I added some code to allow me to do that. Namely resizeHeight and resizeWidth booleans.

I'm using a fork of this repo with these changes and it fixes the issue I was seeing with the height. 
https://www.npmjs.com/package/tg-react-reflex

I wasn't quite sure how to update the documentation to show these new options so I thought I'd ask for your help doing so. 

Thanks! 

Love this repo btw